fix: accidental Markdown usage in Nonce Schema section of permit2 docs
The Nonce Schema section of the permit2 docs refers to the number 2 to
the power of 248; however representing this as `2**248` twice within
the same sentence inadvertently triggering Markdown's mechanism for
bold text.  So escape the `**` characters to avoid this bold
formatting.

Also fix a "this" / "thus" typo, and make a `uint248` monospace for
consistency.